Implement the getFleschScore method in Document.java

+++ src/document/Document.java
@@ -130,7 +130,7 @@ public abstract class Document {
 	/** return the Flesch readability score of this document */
 	public double getFleschScore()
 	{
-	    return 206.835 - ((1.015 * (getNumWords()/getNumSentences())) - (84.6 * (getNumSyllables() / getNumWords())));
+	    return 206.835 - 1.015 * ((double) getNumWords()/getNumSentences()) - 84.6 * ((double) getNumSyllables() / getNumWords());
 	}
